messenger-api-service/pyproject.toml
[project]
name = "telegram-api-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.0",
    "python-telegram-bot>=22.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.40",
    "httpx>=0.28.1",
]
